+++ src/ipc/issueActions.ts
@@ -129,6 +129,12 @@ export interface OpenStartWorkPageAction extends Action {
     issue: MinimalIssue<DetailedSiteInfo>;
 }
 
+export interface CloneIssueAction extends Action {
+    action: 'cloneIssue';
+    site: DetailedSiteInfo;
+    issueData: any;
+}
+
 export interface WorklogData {
     comment: string;
     started: string;
@@ -296,6 +302,10 @@ export function isStartWork(a: Action): a is StartWorkAction {
 
 export function isOpenStartWorkPageAction(a: Action): a is OpenStartWorkPageAction {
     return (<OpenStartWorkPageAction>a).issue !== undefined;
+}
+
+export function isCloneIssue(a: Action): a is CloneIssueAction {
+    return a && a.action === 'cloneIssue';
 }
 
 export function isUpdateAiSettings(a: Action): a is UpdateAiSettingsAction {